# Improve line wrapping performance