--- a/scene/gui/text_edit.cpp
+++ b/scene/gui/text_edit.cpp
@@ -165,15 +165,7 @@ int TextEdit::Text::get_line_wrap_amount(int p_line) const {
 }
 
 Vector<Vector2i> TextEdit::Text::get_line_wrap_ranges(int p_line) const {
-	Vector<Vector2i> ret;
-	ERR_FAIL_INDEX_V(p_line, text.size(), ret);
-
-	Ref<TextParagraph> data_buf = text[p_line].data_buf;
-	int line_count = data_buf->get_line_count();
-	for (int i = 0; i < line_count; i++) {
-		ret.push_back(data_buf->get_line_range(i));
-	}
-	return ret;
+	return text[p_line].data_buf->get_line_ranges();
 }
 
 const Ref<TextParagraph> TextEdit::Text::get_line_data(int p_line) const {