This PR's implementation works, but it's more so meant to start a discussion than to actually be finished. I figure there are most likely better ways to go about this.

In a nutshell, very long lines are really harsh on TextEdit when wrapping is enabled. I made this benchmark in 4.4.beta1 for `get_rect_at_line_column()`, an example of a method that's bottlenecked by this:

- 64 B -> 0.01ms
- 350 B -> 0.03ms
- 1.5 kB -> 0.15ms
- 9 kB -> 3ms
- 62 kB -> 65ms

Not a very tight benchmark, as the amount of linewraps matters more, but it's roughly every 50 characters.

In development builds, with the 9 kB line, I got 7.8ms; with this PR, I get 0.8ms and I don't notice any regressions.

If I gauge correctly, the issue is that `_shape_lines()` in TextParagraph is called numerous times without anything that would actually make a difference. Would caching make more sense then?