@MewPurPur 2025-01-19
What I missed is that it seems like the `TS->shaped_text_is_ready()` checks at the beginning of the line shaping are what's moderately expensive and what's actually ran a lot of times