@bruvzg 2025-01-19
A single `shaped_text_is_ready` should be fast, but I guess if it's called for a lot of wrapped lines (which it is), it can be an issue. We probably should do in it the opposite way, and instead of looping over all lines, set a single flag in the unwrapped parent line when wrapped children are invalidated. I'll take a look tomorrow.